Wrote some code to test how the pickling/unpickling affects performance.
Ok, so the pickling code doesn't actually work in this check-in (the
updated dict isn't saved correctly) but it does show how much of an
effect the pickling has. And it's not good. I've actually cut down the
heart beat interval 100 times and it still takes 1.5 times longer to
complete.