Removed the alpha_count code.
I'm going to try another approach. Instead of walking through the training-
file indexing all words in one bigrams-dict and dumping it to file at
regular intervals I'm going to split the bigrams into buckets based on
first letter of the word from the start. It's probably going to take a
lot more run-time to complete a pass but in comparison to first building
the bigrams and then merging and sorting them it's a lot less complex.