Added a test runner and a sample test sentence.
So the idea is to write some code that uses the bigrams index to figure
out which word is missing and add it in the right place in the sentence.
I'll be testing that code using test.txt and orig.txt. Once I get a
decent implementation I'll see if I can generate the test.txt from
orig.txt directly instead.