Added code to actually produce the hashtable with bigrams.
Unfortunately this is a lot slower than the Python version. Running it
on SBCL causes a crash because it runs out of memory in the heap. CCL
behaves better but is even slower. I added the *MAX-LINES* parameter
so that I could make a comparison. 1 000 000 rows takes about five
minutes with CCL and about 45 seconds in Python.