The process-file now spits out a file with source code.
The idea is to create a source file containing all the lisp code to
produce the bigrams hashtable. It's perhaps not the prettiest of
approaches but it works and it looks promising. I'm assuming that I
can scale this to many hundred thousands of lines of code.

In other news, after some investigations it seems that SBCL and CCL
are about four times slower than Python when it comes to IO. So there.
That explains the poor performance in build.lisp but all will be
forgiven and forgotten as long as they turn out to be blazing fast
in the other areas.